inr-hft-sandbox/requirements.txt
cryptofeed==2.4.0
pyyaml==6.0.1
pandas==2.1.4
colorlog==6.8.0
pytest==7.4.3          # dev only
